Start initial work getting OP_SYSOPEN working
Totally broken, but some work on doing this.

TODO: find out why XS is eating my arguments. open() may or may not have
some weird magic going on or it may just be me.

Also use sysopen so we don't have to use backticks to check the contents
of the file in the tests.