Get multiple arguments working
[Why]
Ideally we want to get all the arguments to open, not just the filename.
Also this helps so we can check how many arguments are received (since
filename is only included with 2 or 3 argument open.

[How]
Add extensive notes and reformatting of the XS code.

Have the max arguments be set at 99, and store them into an array. This
allows us to store them all into this array. Then once they are all
stored we can start touching the stack.