Change calling convention and improve safety w/ recursion
[Why]
1. We want this to be a safe pre-hook. We must ensure that we can call
open() inside of our hook without it recursing.

2. It nicer to separate the loading of the module and the add of the
pre-hook.

[How]
1. Check the CvDEPTH of the pre-hook sub. This ensures we can skip the
pre-hook if the pre-hook calls open. In addition, we also die if the
pre-hook is an XS function. In these functions you can't check the
depth, so have it die. If the user really wants this can wrap it in a
Perl sub.